jaegar/backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.15
Django==5.1.4
djangorestframework==3.15.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
idna==3.10
importlib_metadata==8.5.0
kombu==5.4.2
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-celery==0.50b0
opentelemetry-instrumentation-django==0.50b0
opentelemetry-instrumentation-wsgi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
packaging==24.2
prompt_toolkit==3.0.48
protobuf==5.29.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.0
zipp==3.21.0
